+++ CChromaEditorLibrary/ChromaSDKPluginTypes.cpp
@@ -9,7 +9,7 @@ FChromaSDKColors::FChromaSDKColors()
 FChromaSDKColors::FChromaSDKColors(const FChromaSDKColors& copy)
 {
 	const std::vector<COLORREF>& colors = copy.Colors;
-	int size = colors.size();
+	int size = (int)colors.size();
 	for (int index = 0; index < size; ++index)
 	{
 		const COLORREF& color = colors[index];
@@ -24,7 +24,7 @@ FChromaSDKColorFrame1D::FChromaSDKColorFrame1D()
 FChromaSDKColorFrame1D::FChromaSDKColorFrame1D(const FChromaSDKColorFrame1D& copy)
 {
 	const std::vector<COLORREF>& colors = copy.Colors;
-	int size = colors.size();
+	int size = (int)colors.size();
 	for (int index = 0; index < size; ++index)
 	{
 		const COLORREF& color = colors[index];
@@ -65,7 +65,7 @@ FChromaSDKColorFrame2D::FChromaSDKColorFrame2D(const EChromaSDKDevice2DEnum devi
 FChromaSDKColorFrame2D::FChromaSDKColorFrame2D(const FChromaSDKColorFrame2D& copy)
 {
 	const std::vector<FChromaSDKColors>& colors = copy.Colors;
-	int colorsSize = colors.size();
+	int colorsSize = (int)colors.size();
 	for (int index = 0; index < colorsSize; ++index)
 	{
 		const FChromaSDKColors& color = colors[index];
@@ -73,13 +73,13 @@ FChromaSDKColorFrame2D::FChromaSDKColorFrame2D(const FChromaSDKColorFrame2D& cop
 	}
 	
 	const std::vector<FChromaSDKColors>& keys = copy.Keys;
-	int maxRow = keys.size();
+	int maxRow = (int)keys.size();
 	Keys.clear();
 	for (int i = 0; i < maxRow; ++i)
 	{
 		const FChromaSDKColors& sourceRow = keys[i];
 		FChromaSDKColors row;
-		const int maxColumn = sourceRow.Colors.size();
+		const int maxColumn = (int)sourceRow.Colors.size();
 		for (int j = 0; j < maxColumn; ++j)
 		{
 			row.Colors.push_back(0);
